Refactor out the new marshalling model into its own IMarshallingGeneratorFactory implementation to enable more localized handling of collection element marshalling support with the decorator pattern.
